requirements.txt
requests
pandas
boto3
snowflake-connector-python
great_expectations
apache-airflow
